Misc improvements around compositions
Each composition type results in both an interface and a class that implements the interface. In previous releases, if a composition had any compositions on its own, only the generated class would implement the interfaces for those compositions, but not the generated interfaces. As of this release, the generated interface will now implement the composition interfaces instead.

If a type is a composition, the properties generated for that composition will call a static method in the class for that composition. Since the composition class may be located in a different namespace, this commit will ensure that a using statement is added for the namespace if different from the model namespace.